+++ src/Mailer.php
@@ -6,6 +6,7 @@ use Stampie\Extra\Event\MessageEvent;
 use Stampie\MailerInterface;
 use Stampie\MessageInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface as ContractsEventDispatcherInterface;
 
 /**
  * MailerInterface decorator dispatching events.
@@ -30,7 +31,12 @@ class Mailer implements MailerInterface
     public function send(MessageInterface $message)
     {
         $event = new MessageEvent($message);
-        $this->dispatcher->dispatch(StampieEvents::PRE_SEND, $event);
+
+        if ($this->dispatcher instanceof ContractsEventDispatcherInterface) {
+            $this->dispatcher->dispatch($event, StampieEvents::PRE_SEND);
+        } else {
+            $this->dispatcher->dispatch(StampieEvents::PRE_SEND, $event);
+        }
 
         $this->delegate->send($event->getMessage());
     }